+++ docs/CHANGE_LOG.md
@@ -4,6 +4,8 @@
 
 #### **Binding Options - Custom Triggers:**
 - Added a new binding option called "onRefresh", which states an event that should be triggered when a heatmap is refreshed.
+- Added a new binding option called "onBeforeRender", which states an event that should be triggered before an element is rendered.
+- Added a new binding option called "onRenderComplete", which states an event that should be triggered after an element is rendered.
 
 #### **Public Functions:**
 - Added new public function "removeDate()", which will remove a specific date from a heat map (or decrease the date if more than one exists).